requirements.txt
beautifulsoup4==4.8.0
Django==2.1.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.0.1
django-crispy-forms==1.7.2
django-hitcount==1.3.1
django-multiselectfield==0.1.8
djangorestframework==3.10.3
Pillow==6.1.0
pytz==2019.2
soupsieve==1.9.3
